+++ MANIFEST.in
@@ -1,3 +1,1 @@
-include deepSNIaID/trained_models/*.pth
-include deepSNIaID/trained_models/*.npy
-include deepSNIaID/trained_models/*.json
+include deepSIP/trained_models/*.pth
