+++ deepSIP/__init__.py
@@ -1,0 +1,4 @@
+from deepSIP.model import deepSIP
+
+__all__ = ['deepSIP']
+__version__ = '0.1.dev0'
