+++ deepSIP/architecture.py
