--- deepSNIaID/__init__.py
@@ -1,4 +1,0 @@
-from deepSNIaID.model import deepSNIaID
-
-__all__ = ['deepSNIaID']
-__version__ = '0.1.dev0'
